Abstract data table component
